refactor(encrypt & decrypt): rename internal encrypt & decrypt and add -doc
Internal used encrypt and decrypt are now renamed to encrypt-doc and decrypt-doc. Because they are used for the documents.
