+++ lib/decrypt-doc.js
