+++ lib/encrypt-doc.js
