+++ lib/helpers/create-reset-keys.js
@@ -5,7 +5,7 @@ var Promise = require('lie')
 var randomBytes = require('randombytes')
 
 var createKey = require('../create-key')
-var encrypt = require('../encrypt')
+var encrypt = require('../encrypt-doc')
 
 module.exports = createResetKeys
