+++ lib/helpers/decrypt-one.js
@@ -3,7 +3,7 @@
 var Promise = require('lie')
 var PouchDBErrors = require('pouchdb-errors')
 
-var decrypt = require('../decrypt')
+var decrypt = require('../decrypt-doc')
 
 module.exports = decryptOne
