+++ lib/helpers/encrypt-one.js
@@ -3,7 +3,7 @@
 var Promise = require('lie')
 var PouchDBErrors = require('pouchdb-errors')
 
-var encrypt = require('../encrypt')
+var encrypt = require('../encrypt-doc')
 
 module.exports = encryptOne
