+++ tests/integration/change-password.js
@@ -6,7 +6,7 @@ const pouchdbErrors = require('pouchdb-errors')
 
 const createCryptoStore = require('../utils/createCryptoStore')
 const createKey = require('../../lib/create-key')
-const decrypt = require('../../lib/decrypt')
+const decrypt = require('../../lib/decrypt-doc')
 
 test('cryptoStore.changePassword should only exist on the root api', async t => {
   t.plan(3)
